Refactor: Enforce node affinity for essential workloads
This commit refactors the node affinity configuration for several Kubernetes deployments to ensure that essential workloads are scheduled on nodes with the `la1r.workload/essential` label.

The previous configuration used `preferredDuringSchedulingIgnoredDuringExecution`, which only *preferred* nodes with the specified label. This has been changed to `requiredDuringSchedulingIgnoredDuringExecution` with `nodeSelectorTerms` to *require* that the workloads are scheduled on nodes with the `la1r.workload/essential` label.

Additionally, the `priorityClassName: essential` has been removed from several deployments, as the node affinity now guarantees that the workloads will be scheduled on the appropriate nodes.